requirements.txt
absl-py~=1.2.0
protobuf<3.20,>=3.9.2
tensorflow~=2.10.0
tensorflow-datasets~=4.6.0
jax
chex
optax
scikit-learn
ml-collections==0.1.1
dill<0.3.2,>=0.3.1.1
Pillow
pyquaternion
ncompress
xmltodict
pandas
apache-beam
kaggle
